chore(plugin): add simple logging implementation
We wrap hclog as this seems to be the recommended library to use with
the go-plugin architecture. We only support logging to file as any other
destination doesn't end up where we would expect it to.

Also includes some cleanup of the error handling as this commit touched
that already anyway. Next commit will include fixes for the SQS error
handling.

Closes: #217