fix(plugin): publish large message payloads to S3
To have some leeway we upload to S3 at ~100KB rather than the exact
limit of SQS. This means we should realistically always be able to
publish the upload message body with the 'normal' message attributes.

Also adds a bunch of extra logs to make debugging things easier when it
comes to SQS related issues.

Closes: #247