chore(plugin): add message filtering capability
Optionally allow configuring a list of message types that are allowed to
be published on the queue. Went for an allowlist as we'll generally use
it when we want to reindex historical state and only publish updates for
the new messages.

For the same reason I went for a simple slice of strings as the data
structure; the list should never be long enough to matter in terms of
performance.

Closes: #258