feat(x/tally): store the exit message in the result for tallyvm errors
The tallyvm can run into errors outside the actual VM, while resolving
host bindings for example. This leads to only the exit_info.message to
be set and not the result. We write the message to the result in this
scenario so it's easier for a developer to see what kind of error their
program ran into.

Closes: #454