chore(x/tally): update endblock error handling
Main difference is that we now allow the node to panic if we are unable
to write to the store. This prevents us from continuing on invalid state
should something go wrong there. It should never happen due to inputs as
those get validated at mutliple points throughout the code.

We also write an empty result to the store in case there are
encoding/decoding errors for fields that we are in control of.