fix(abci): handle invalid vote extensions in late votes
Since votes that arrive before the timeout but after the block has been
delivered to the application are not verified with the
verifyVoteExtension handler we need to account for potentially invalid
votes in the prepareProposalHandler. Since we at least are guaranteed to
receive verified votes for over 2/3rds of voting power we can simply
ignore any vote that was invalid.