fix(plugin): add additional validity check for deserialised Txs
For some reason the injected vote extension bytes sometimes do get
serialised into the Tx object even though they shouldn't. As a
workaround we verify that when there is no error that the Tx body is not
nil. If it is nil we do go through the vote injection flow and error if
it's also not a valid vote extension.