feat(x/batching): pruning of data results and their batch assignments
We resort to naive implementation of data results pruning because there
is no mapping to data result objects from batch number or data result ID,
which is only data available to us when we prune batches. In this
implementation we go through first `MaxDataResultsToCheckForPrune` items
in the store and delete those whose associated batches have been pruned.